+++ internal/database/database.go
@@ -5,7 +5,7 @@ import (
 	"os"
 
 	_ "github.com/lib/pq"
-	"github.com/nstoker-clixifix/find_script/internal/logger"
+	"github.com/sirupsen/logrus"
 )
 
 var DB *sql.DB
@@ -13,20 +13,20 @@ var DB *sql.DB
 func Connect() {
 	database_url := os.Getenv("DATABASE_URL")
 	if database_url == "" {
-		logger.Log.Fatal("environment variable DATABASE_URL not set")
+		logrus.Fatal("environment variable DATABASE_URL not set")
 	}
 
-	logger.Log.Info("connecting to database")
+	logrus.Info("connecting to database")
 	db, err := sql.Open("postgres", database_url)
 	if err != nil {
-		logger.Log.Fatal(err.Error())
+		logrus.Fatal(err.Error())
 	}
 	if db == nil {
-		logger.Log.Fatal("did not attach to database")
+		logrus.Fatal("did not attach to database")
 	}
 
 	if err = db.Ping(); err != nil {
-		logger.Log.Fatal(err)
+		logrus.Fatal(err)
 	}
 
 	DB = db